gen: Fun-, Hybrid-

+++ tla-types/src/main/scala/at/forsyte/apalache/tla/typecomp/subbuilder/LiteralAndNameBuilder.scala
@@ -16,7 +16,7 @@ trait LiteralAndNameBuilder {
   private val unsafeBuilder = new UnsafeLiteralAndNameBuilder
 
   /** {{{i : Int}}} */
-  def int(i: BigInt): TBuilderInstruction = unsafeBuilder.int(i).point[TBuilderInternalState]
+  def int(i: Int): TBuilderInstruction = unsafeBuilder.int(i).point[TBuilderInternalState]
 
   /**
    * {{{s : Str}}}
